Fix M3 oil loss: cast mask targets to int64 (Long)
Albumentations resize casts segmentation masks to int32, but F.cross_entropy and
F.one_hot require int64. Cast in OilSpillDataset.__getitem__ and defensively in
DiceFocalLoss.forward. Surfaced by the Colab smoke test (would have crashed on the
first real epoch).

+++ src/data/loaders.py
@@ -190,7 +190,8 @@ class OilSpillDataset(Dataset):
         if self.transform:
             augmented = self.transform(image=img, mask=mask)
             img = augmented["image"]
-            mask = augmented["mask"]
+            # albumentations resize casts masks to int32 — CrossEntropy/one_hot need int64
+            mask = augmented["mask"].long()
         else:
             # default resize + to tensor
             img = torch.from_numpy(img.transpose(2, 0, 1)).float()
